+++ client/src/components/dashboard/Classes/showTasks.js
@@ -9,16 +9,18 @@ import { removeTask } from '../../../redux/action';
 import moment from 'moment';
 
 const mapStateToProps = state => {
-    return { classes: state.classes, classIndex: state.classIndex, classID: state.classID };
+    return { classes: state.classes, classIndex: state.classIndex, classID: state.classID, isGuess: state.isGuess };
 };
 
 class ShowTasks extends React.Component{
 
     handleOnClick = id => {
-        fetch(`/api/class/task?classID=${this.props.classID}&taskID=${id}`, {
-            method: 'DELETE',
-            credentials: 'include',
-        });
+        if(!this.props.isGuess){
+            fetch(`/api/class/task?classID=${this.props.classID}&taskID=${id}`, {
+                method: 'DELETE',
+                credentials: 'include',
+            });
+        }
         this.props.dispatch(removeTask(id));
     }
